Extract harness code into its own module.
To keep the test code clean we pull out more general harness stuff into
a seperate file.
